WIP: refactor all S3 operations out into one method.
This is far from what I think the final product should be. I'm grasping
for the correct abstraction, and this is a middle point to funnel all
current S3 operations through one method.

The unreadable complexity of the resulting frankenmethod plus the
perplexed comment I left towards the bottom should indicate that
many changes must still be made.